feat <김상현 #9> create response dto
